NET-2: moved packet to phase 2 packet file
